[UPDATE] UI 구조 업데이트 (분절)
