+++ Assets/Prefabs/StatsUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f672b60c3e5bf2449abb65a3623e1b95
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
