+++ Assets/Scripts/UI/Buttons/ExitBtn.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public class ExitManager : MonoBehaviour
+{
+    public void OnExitButtonClicked()
+    {
+        // 게임 종료 로직
+        Debug.Log("Game is exiting...");
+        Application.Quit();
+    }
+}
