+++ Assets/Scripts/UI/Buttons/ExitBtn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4bdd0ed4d98a434fbbc3d0878390eab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
