+++ Assets/Scripts/UI/Buttons/OptionsBtn.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+public class OptionsBtn : MonoBehaviour
+{
+    public GameObject optionsMenu;
+
+    void Start()
+    {
+    }
+
+    public void OnOptionsButtonClicked()
+    {
+        optionsMenu.SetActive(true);
+    }
+
+}
