+++ Assets/Scripts/UI/Buttons/StartBtn.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a2d00c5e960c2ad4a950ac467e250f66
+guid: e4282d16def726b4389083537299f327
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
