+++ Assets/Scripts/UI/Menu.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 03bae987fa6547d409553cf59f0d5f91
-PrefabImporter:
+guid: d3f2230a6c2e16d4e8d14879f6b448a7
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
