+++ Assets/Scripts/UI/Menu/AudioSettingMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d303bcba88b2ed46aa16cc5afa08954
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
