+++ Assets/Scripts/UI/Menu/CrosshairSettingMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aca753f4dd55294449fac867d15f1450
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
