+++ Assets/Scripts/UI/Menu/MouseSettingMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04338c939cc75f64a886d5bdf080f24f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
