+++ Assets/Scripts/UI/Menu/OptionMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c58cb682754e9554bb8acfd1a33ea995
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
